move `NewCookie` and `RandNonZeroUint64` out
