+++ pkg/security/probe/process_resolver.go
@@ -273,7 +273,7 @@ func (p *ProcessResolver) DequeueExited() {
 func (p *ProcessResolver) NewProcessCacheEntry(pidContext model.PIDContext) *model.ProcessCacheEntry {
 	entry := p.processCacheEntryPool.Get()
 	entry.PIDContext = pidContext
-	entry.Cookie = eval.NewCookie()
+	entry.Cookie = utils.NewCookie()
 	return entry
 }
