+++ pkg/security/secl/compiler/eval/rand.go
@@ -24,18 +24,3 @@ func RandString(n int) string {
 func init() {
 	rand.Seed(time.Now().UnixNano())
 }
-
-// NewCookie returns a new random cookie
-func NewCookie() uint32 {
-	return rand.Uint32()
-}
-
-// RandNonZeroUint64 returns a new non-zero uint64
-func RandNonZeroUint64() uint64 {
-	for {
-		value := rand.Uint64()
-		if value != 0 {
-			return value
-		}
-	}
-}
